Extract react window and render rows to separate file
