+++ src/components/list/FixedSizeList.js
@@ -1,0 +1,57 @@
+import { ChevronRight, Star } from '@styled-icons/boxicons-solid'
+import { FixedSizeList as List } from 'react-window'
+
+import { theme } from '../ui/GlobalStyle'
+import ListEntry from './ListEntry'
+
+const convertMetersToMiles = (meters) => {
+  let miles = (meters * 0.000621371192).toString()
+  miles = miles.slice(0, miles.indexOf('.') + 3)
+  return miles
+}
+
+const FixedSizeList = ({
+  height,
+  width,
+  itemSize,
+  itemCount,
+  isItemLoaded,
+  locations,
+  handleListEntryClick,
+  ...props
+}) => {
+  const renderRow = ({ index, style }) => {
+    let row
+    if (!isItemLoaded(index)) {
+      row = <ListEntry primaryText="Loading..." style={style} />
+    } else {
+      const location = locations[index]
+      row = (
+        <ListEntry
+          key={location.id}
+          leftIcons={<Star size="16" />}
+          rightIcons={<ChevronRight size="16" color={theme.blue} />}
+          primaryText={location.type_names[0]}
+          secondaryText={`${convertMetersToMiles(location.distance)} miles`}
+          onClick={() => handleListEntryClick(location.id)}
+          style={style}
+        />
+      )
+    }
+    return row
+  }
+
+  return (
+    <List
+      height={height}
+      width={width}
+      itemSize={itemSize}
+      itemCount={itemCount}
+      {...props}
+    >
+      {renderRow}
+    </List>
+  )
+}
+
+export default FixedSizeList
