+++ src/components/mobile/SettingsPage.js
@@ -3,10 +3,10 @@ import { Cog, Flag, Star } from '@styled-icons/boxicons-solid'
 import { useEffect, useState } from 'react'
 
 import { getTypes } from '../../utils/api'
+import ListEntry from '../list/ListEntry'
 import Button from '../ui/Button'
 import CaptionInput from '../ui/CaptionInput'
 import Input from '../ui/Input'
-import ListEntry from '../ui/ListEntry'
 import Modal from '../ui/Modal'
 import ProgressBar from '../ui/ProgressBar'
 import { Select } from '../ui/Select'
