--- src/components/ui/List.js
@@ -1,76 +1,0 @@
-import { ChevronRight, Star } from '@styled-icons/boxicons-solid'
-import AutoSizer from 'react-virtualized-auto-sizer'
-import { FixedSizeList } from 'react-window'
-import InfiniteLoader from 'react-window-infinite-loader'
-
-import { theme } from '../ui/GlobalStyle'
-import ListEntry from '../ui/ListEntry'
-
-const convertMetersToMiles = (meters) => {
-  let miles = (meters * 0.000621371192).toString()
-  miles = miles.slice(0, miles.indexOf('.') + 3)
-  return miles
-}
-
-const List = ({
-  locations,
-  handleListEntryClick,
-  loadNextPage,
-  hasMoreItems,
-  isNextPageLoading,
-}) => {
-  const itemCount = hasMoreItems ? locations.length + 1 : locations.length
-
-  // eslint-disable-next-line no-empty-function
-  const loadMoreItems = isNextPageLoading ? () => {} : loadNextPage
-
-  const isItemLoaded = (index) => !hasMoreItems || index < locations.length
-
-  const renderRow = ({ index, style }) => {
-    let row
-    if (!isItemLoaded(index)) {
-      row = <ListEntry primaryText="Loading..." style={style} />
-    } else {
-      const location = locations[index]
-      row = (
-        <ListEntry
-          key={location.id}
-          leftIcons={<Star size="16" />}
-          rightIcons={<ChevronRight size="16" color={theme.blue} />}
-          primaryText={location.type_names[0]}
-          secondaryText={`${convertMetersToMiles(location.distance)} miles`}
-          onClick={() => handleListEntryClick(location.id)}
-          style={style}
-        />
-      )
-    }
-    return row
-  }
-
-  return (
-    <InfiniteLoader
-      isItemLoaded={isItemLoaded}
-      itemCount={itemCount}
-      loadMoreItems={loadMoreItems}
-    >
-      {({ onItemsRendered, ref }) => (
-        <AutoSizer>
-          {({ height, width }) => (
-            <FixedSizeList
-              height={height}
-              width={width}
-              itemSize={57}
-              itemCount={itemCount}
-              onItemsRendered={onItemsRendered}
-              ref={ref}
-            >
-              {renderRow}
-            </FixedSizeList>
-          )}
-        </AutoSizer>
-      )}
-    </InfiniteLoader>
-  )
-}
-
-export default List
